status update of extended notifications
