+++ src/scss/notification.scss
@@ -2,14 +2,13 @@
 
 as24-notification {
   overflow: hidden;
-  position: absolute;
   width: 100%;
-  height: 0;
+  display: none;
   z-index: 9999;
   transition: height 0.4s ease;
 
   &.show {
-    height: 71px;
+    display: block;
   }
 
   @each $name, $value in $types {
@@ -56,10 +55,8 @@ as24-notification {
 }
 
 .sc-notification-container {
-  border: 1px solid black;
   z-index: 9999;
   overflow: hidden;
   position: absolute;
   width: 100%;
-  height: 100px;
 }
